feat: added scrape engine with request and parse functionality
Added new file scrape_engine.py to capture functions relating to the webpage scraping functionality. Currently contains functions for request and parse, as well as tabulating the data for visual representation in gui, email, etc.

Additional considerations for error handling will need to be included in upcoming commits, as the error handling convention is formalized.